+++ frontend/public/js/textViewer.js
@@ -1,9 +1,0 @@
-import { $textViewer } from "./components.js";
-import { clearImage, loadImage } from "./imageViewer.js";
-$textViewer.on("mouseover", async (e) => {
-  if (e.target.tagName === "A") {
-    loadImage(e.target.href);
-  } else {
-    clearImage();
-  }
-});