+++ frontend/public/js/uploadImage.js
@@ -1,15 +1,34 @@
-import { $input, $inputForm, $uploadImage } from "./components.js";
+import { $input, $inputForm } from "./components.js";
 import saveFile from "./saveFile.js";
 
 if (!$input.length) {
-  throw new Error("No input element found");
+  console.error("No input element found");
 }
 if (!$inputForm.length) {
-  throw new Error("No input form found");
+  console.error("No input form found");
 }
-if (!$uploadImage.length) {
-  throw new Error("No upload image button found");
-}
+
+const handleUploadImage = async () => {
+  var selection = window.getSelection();
+  var selectedText = selection.toString();
+  if (selectedText && !/^\s+$/.test(selectedText)) {
+    var range = selection.getRangeAt(0);
+    if (range.startContainer === range.endContainer) {
+      uploadImage().then((response) => {
+        var $anchor = $("<a>", {
+          href: response,
+          html: selectedText,
+          onclick: "event.preventDefault()",
+          display: "inline-block",
+        });
+        $anchor.data("id", response);
+        range.deleteContents();
+        range.insertNode($anchor[0]);
+        saveFile();
+      });
+    }
+  }
+};
 
 async function uploadImage() {
   return new Promise((resolve, reject) => {
@@ -44,25 +63,12 @@ async function uploadImage() {
   });
 }
 
-$uploadImage.on("click", async () => {
-  var selection = window.getSelection();
-  var selectedText = selection.toString();
-  if (selectedText && !/^\s+$/.test(selectedText)) {
-    var range = selection.getRangeAt(0);
-    if (range.startContainer === range.endContainer) {
-      var $anchor = $("<a>", {
-        href: "/res/loading-sm.gif",
-        html: selectedText,
-        onclick: "event.preventDefault()",
-        display: "inline-block",
-      });
-      range.deleteContents();
-      range.insertNode($anchor[0]);
-
-      var response = await uploadImage();
-      $anchor.attr("href", response);
-      $anchor.data("id", response);
-      saveFile();
-    }
-  }
-});
+export var uploadImageButton = () => {
+  var button = $("<button>", {
+    id: "upload-image",
+    text: "Upload Image",
+    class: "dropdown-item",
+  });
+  button.on("click", handleUploadImage);
+  return button;
+};